Move benchmarks to separate file, add new benchmark tests
